refactor(headers): give integer an arithmetic core and free posit's core of I/O (#1334)
Phase N of the header-layering epic -- the two items it named as prerequisites for
the per-type rollout, plus the blocktriple forward declaration.

  header                                     lines  headers   I/O
  posit/core.hpp            (before)        107282      437     5
  posit/core.hpp            (after)          77433      369     0
  integer/integer.hpp       (before)        103416      420     5
  integer/core.hpp          (new)            48324      217     0
  utility/decimal_to_binary (before)        103826      422     5
  utility/decimal_to_binary (after)          57365      265     0

integer gains core.hpp and iostream.hpp beside its existing manipulators.hpp, and
decimal_to_binary.hpp -- reached from posit, cfloat, fixpnt, dd and qd -- depends on
that core rather than the umbrella, so it stops putting all five I/O-family headers
into every one of those cores. The umbrella keeps its name and its contents, so no
caller changes (compatibility option (a), as decided in the epic).

What left integer's core:
  - convert_to_decimal_string/to_string/to_binary/to_native -> manipulators.hpp
  - convert_to_string/operator<</operator>>                 -> iostream.hpp (new)
  - support/decimal.hpp (80048 lines, 4 I/O) followed its only caller
  - internal/blockbinary/manipulators.hpp (50808 lines, 4 I/O) -- was never used
  - native/integers.hpp (44662 lines, 3 I/O) -- the core needed only nlz(), which
    moves to native/integer_core.hpp, the same split #1400 gave ieee754
  - an unused <vector> (26261 lines)
  - 9 std::cerr/std::cout diagnostics -> fprintf(stderr, ...), the Phase 0 idiom

posit's three field headers each carried operator<<, operator>> and to_string, so its
arithmetic core pulled <sstream>/<ostream>/<istream> for debug printers it never calls;
posit_fraction.hpp reached the stream operators only through an unrelated <regex> that
happened to be in the graph. That text layer is now posit_fields_io.hpp, kept separate
from posit/iostream.hpp because posit_impl.hpp needs it under POSIT_TRACE_ENABLED and
posit/iostream.hpp includes posit_impl.hpp -- folding them would be circular.

blocktriple_fwd.hpp carries the BlockTripleOperator enum and the class declaration with
its default template arguments (a template may declare defaults only once), so a header
that merely names blocktriple need not pay ~70000 lines for it. integer::normalize() is
a member template, so its body instantiates only where it is called and the core never
needed the definition. posit cannot use this: its +=, *= and /= construct blocktriple
locals, so its core stays at blocktriple's own 69758 lines.

Also removes two orphaned files under internal/blocktriple/ that no CMake glob picked
up and that had not compiled in a long time.

Verified: five gates -- integer core alone, integer core+manipulators, posit core alone,
posit core+manipulators, and posit's trace path -- each compiled AND run on gcc and
clang, per the epic's rule that the umbrella building proves nothing. Full sweep of
every api.cpp plus the blocktriple, blockbinary, blockfraction, blocksignificand,
integer, einteger, posit and quire suites: 162 pass, 0 fail, 0 warnings. ASCII clean.

integer/core.hpp meets both acceptance criteria (0 I/O, under 45000 is 48324, within
7%). posit/core.hpp meets criterion 1; criterion 2 is bounded by blocktriple and needs
a separate decision rather than a moved bar.

BEHAVIOUR-CHANGE: integer<>::showLimbs() prints the full limb encoding instead of one
bit per limb -- it passed the bit width into to_binary's nibble-marker parameter and
`true` into its nbits parameter; positRegime's "Regime needs to expand" diagnostic now
goes to stderr rather than stdout
